COMMIT FUNCAO

+++ SysGeSeApp2024/Interfaces/IRepository.cs
@@ -4,7 +4,7 @@ namespace SysGeSeApp2024.Interfaces
 {
     public interface IRepository<TEntity> : IDisposable where TEntity : Entity
     {
-        Task<TEntity> ObterPorId(long id);
+        Task<TEntity> ObterPorId(int id);
 
         Task<List<TEntity>> ObterTodos();
         Task Adicionar(TEntity entity);
